Eliminate 'conversion may change sign of result' clang warnings in libgc
Issue #558 (bdwgc).

* allchblk.c (GC_merge_unmapped, GC_get_first_part, GC_allochblk_nth):
Cast ~WAS_UNMAPPED and ~FREE_BLK to unsigned char.
* alloc.c (min_bytes_allocd): Cast pointer difference to word.
* dbg_mlc.c (GC_debug_free, GC_debug_free_inner, GC_debug_realloc):
Likewise.
* headers.c (GC_install_counts): Likewise.
* include/private/gc_pmark.h (GC_push_contents_hdr): Likewise.
* mark.c (GC_set_mark_bit, GC_clear_mark_bit, GC_is_marked): Likewise.
* mark_rts.c (GC_compute_root_size, GC_add_roots_inner,
GC_remove_roots_inner): Likewise.
* os_dep.c (GC_remap): Likewise.
* ptr_chck.c (GC_is_visible): Likewise.
* alloc.c (GC_adj_bytes_allocd): Cast GC_bytes_allocd and
GC_bytes_finalized to signed_word.
* alloc.c (GC_clear_fl_marks): Cast sz to signed_word.
* alloc.c [USE_MUNMAP] (GC_try_to_collect_general): Change type of
old_unmap_threshold local variable to unsigned int.
* backgraph.c (reset_back_edge): Cast ~RETAIN to unsigned short.
* backgraph.c (backwards_height): Cast be->height to word; cast
p_height to signed_word.
* dbg_mlc.c (GC_get_back_ptr_info): Cast pointer difference to size_t.
* fnlz_mlc.c (GC_register_disclaim_proc_inner): Change type of kind
argument to unsigned int.
* fnlz_mlc.c (GC_register_disclaim_proc): Cast kind argument to
unsigned int.
* fnlz_mlc.c (GC_finalized_malloc): Cast GC_finalized_kind to int.
* gcj_mlc.c (GC_init_gcj_malloc): Cast result of GC_new_kind_inner()
to int.
* typd_mlc.c (GC_init_explicit_typing): Likewise.
* headers.c (GC_prev_block): Cast j local variable to word.
* include/gc/gc_mark.h (GC_DS_TAGS): Change 1 to 1U.
* include/gc/gc_mark.h (GC_MAKE_PROC): Cast proc_index to unsigned int;
cast GC_DS_PROC to GC_word.
* include/gc/gc_typed.h (GC_get_bit, GC_set_bit): Add comment about
index argument.
* mallocx.c (GC_generic_malloc_many): Change type of my_bytes_allocd
local variable to word; cast my_bytes_allocd_tmp to word; cast
my_bytes_allocd to signed_word.
* mark.c (GC_mark_from): Cast WORDS_TO_BYTES(WORDSZ/2) to signed_word.
* misc.c (GC_init): Cast time_limit to unsigned long; cast
unmap_threshold to unsigned int.
* misc.c (GC_write): Cast result (len) to int; cast bytes_written to
size_t; cast result of write() to int.
* pthread_support.c (do_blocking_leave, GC_call_with_gc_active,
GC_register_my_thread): Cast ~DO_BLOCKING and ~FINISHED to unsigned
char.
* reclaim.c (GC_clear_block, GC_reclaim_clear, GC_reclaim_uninit,
GC_disclaim_and_reclaim, GC_reclaim_generic): Change name of count
argument to pcount and change its type to word.
* include/private/gc_priv.h (GC_generic_malloc_many): Likewise.
* reclaim.c (GC_reclaim_small_nonempty_block,
GC_disclaim_and_reclaim_or_free_small_block): Cast &GC_bytes_found
to word*.
* reclaim.c (GC_disclaim_and_reclaim_or_free_small_block,
GC_reclaim_block): Cast HBLKSIZE and sz to signed_word.
* typd_mlc.c (GC_typed_mark_proc_index, GC_array_mark_proc_index):
Change type from int to unsigned int.
* include/private/gc_priv.h (GC_arrays._finalized_kind): Likewise.
* typd_mlc.c (GC_add_ext_descriptor): Cast GC_avail_descr to
signed_word; cast result local variable to size_t; add assertion that
result is non-negative.
* typd_mlc.c (GC_make_descriptor): Cast last_set_bit, i local variables
to word.
* win32_threads.c (GC_start_world): Cast ~IS_SUSPENDED to unsigned
char.

+++ fnlz_mlc.c
@@ -52,10 +52,11 @@ STATIC int GC_CALLBACK GC_finalized_disclaim(void *obj)
     return 0;
 }
 
-STATIC void GC_register_disclaim_proc_inner(int kind, GC_disclaim_proc proc,
+STATIC void GC_register_disclaim_proc_inner(unsigned kind,
+                                            GC_disclaim_proc proc,
                                             GC_bool mark_unconditionally)
 {
-    GC_ASSERT((unsigned)kind < MAXOBJKINDS);
+    GC_ASSERT(kind < MAXOBJKINDS);
     if (EXPECT(GC_find_leak, FALSE)) return;
 
     GC_obj_kinds[kind].ok_disclaim_proc = proc;
@@ -95,7 +96,8 @@ GC_API void GC_CALL GC_register_disclaim_proc(int kind, GC_disclaim_proc proc,
                                               int mark_unconditionally)
 {
     LOCK();
-    GC_register_disclaim_proc_inner(kind, proc, (GC_bool)mark_unconditionally);
+    GC_register_disclaim_proc_inner((unsigned)kind, proc,
+                                    (GC_bool)mark_unconditionally);
     UNLOCK();
 }
 
@@ -107,7 +109,8 @@ GC_API GC_ATTR_MALLOC void * GC_CALL GC_finalized_malloc(size_t lb,
     GC_ASSERT(GC_finalized_kind != 0);
     GC_ASSERT(NONNULL_ARG_NOT_NULL(fclos));
     GC_ASSERT(((word)fclos & FINALIZER_CLOSURE_FLAG) == 0);
-    op = GC_malloc_kind(SIZET_SAT_ADD(lb, sizeof(word)), GC_finalized_kind);
+    op = GC_malloc_kind(SIZET_SAT_ADD(lb, sizeof(word)),
+                        (int)GC_finalized_kind);
     if (EXPECT(NULL == op, FALSE))
         return NULL;
 #   ifdef AO_HAVE_store
